+++ RecommenderSystem/src/test/ItemBasedTest.java
@@ -10,10 +10,11 @@ public class ItemBasedTest {
 
 	private final double epsilon = 0.001;
 
-	int[][] trainData = { { 1, 1, 5 }, { 1, 2, 3 }, { 1, 3, 4 }, { 2, 1, 4 },
-			{ 2, 2, 2 }, { 2, 3, 4 }, { 2, 5, 3 }, { 3, 2, 2 }, { 3, 4, 4 },
-			{ 3, 5, 3 } };
-	int[][] testData = { { 1, 5, 4 }, { 2, 4, 4 }, { 3, 1, 4 }, { 1, 1582, 1 } };
+	double[][] trainData = { { 1, 1, 5 }, { 1, 2, 3 }, { 1, 3, 4 },
+			{ 2, 1, 4 }, { 2, 2, 2 }, { 2, 3, 4 }, { 2, 5, 3 }, { 3, 2, 2 },
+			{ 3, 4, 4 }, { 3, 5, 3 } };
+	double[][] testData = { { 1, 5, 4 }, { 2, 4, 4 }, { 3, 1, 4 },
+			{ 1, 1582, 1 } };
 
 	@Test
 	public void testItemBasedPredictor() {
@@ -39,7 +40,8 @@ public class ItemBasedTest {
 		double[] predictions = new double[N];
 
 		for (int i = 0; i < N; i++) {
-			predictions[i] = p.predict(testData[i][0], testData[i][1]);
+			predictions[i] = p.predict((int) testData[i][0],
+					(int) testData[i][1]);
 		}
 
 		// Test predictions
